+++ tools/update_ui_icons.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+python auto_atlas.py ui_icons_structure.json > ../core/assets/ui/ui_icons.atlas