+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Only one menu can be visible on the screen at any time
 - Menus are now hidden automatically after selecting an action
 - Undo now works properly after deleting a terrain patch
+- auto_atlas.py tool now accepts a structure.json file that describes the atlas 
 
 ## [1.3] 2015-04-12
 ### Added